+++ contaperu/asiento/__init__.py
@@ -59,7 +59,7 @@ from .resolucion import (comprobantes_sin_centro, comprobantes_sin_cuenta, con_r
 from .huella import huella
 from .imputacion import Imputacion, Parte
 from .lineas import LineaDiario
-from .indice import Cabecera, ComprobanteDelAsiento
+from .indice import Cabecera, ComprobanteDelAsiento, exportacion_de
 from .motor import (CENTRO_EN_ANEXO, ROLES, cabecera_de, constancia_de, glosa_de, lineas_del_comprobante,
                     lineas_del_libro, lineas_e_indice_del_libro, serie_numero_de)
 
@@ -75,5 +75,5 @@ __all__ = [
     "repartos_que_no_cuadran", "sigla_documento", "sub_diario", "sub_diarios_presentes", "tiene_detraccion",
     "tipos_sin_sigla", "huella", "Imputacion", "Parte", "LineaDiario", "CENTRO_EN_ANEXO", "ROLES", "glosa_de",
     "lineas_del_comprobante", "lineas_del_libro", "lineas_e_indice_del_libro", "cabecera_de", "Cabecera",
-    "ComprobanteDelAsiento", "MONEDAS_CODIGO", "constancia_de", "serie_numero_de",
+    "ComprobanteDelAsiento", "MONEDAS_CODIGO", "constancia_de", "exportacion_de", "serie_numero_de",
 ]